+++ src/features/blog/components/index.ts
@@ -1,5 +1,5 @@
 export { BlogCard, CATEGORY_BADGE } from "@/components";
 export { BlogFeaturedCard } from "./BlogFeaturedCard";
-export { BlogCategoryTabs, BLOG_CATEGORY_TABS } from "./BlogCategoryTabs";
+export { BlogCategoryTabs } from "./BlogCategoryTabs";
 export { BlogListView } from "./BlogListView";
 export { BlogPostView } from "./BlogPostView";